requirements.txt
psycopg2
dj-database-url
gunicorn
djangorestframework
djangorestframework-simplejwt

# certifi==2022.9.24
# crispy-bootstrap5==0.7
dj-database-url==1.3.0
# Django==4.1.7
# django-crispy-forms==2.0

# djangorestframework==3.14.0
# djangorestframework-simplejwt==5.2.2
# djangorestframework-jwt==1.11.0


# djangorestframework-jwt==1.11.0
# djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
# numpy==1.24.0
# pandas==1.5.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
# PyJWT==1.7.1
python-dateutil==2.8.2
# pytz==2022.7
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
typing_extensions==4.5.0
absl-py==1.4.0
anyio==3.6.2
asgiref==3.6.0
astunparse==1.6.3
attrs==22.2.0
bleach==6.0.0
CacheControl==0.12.11
cachetools==5.3.0
# certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
dj-database-url==1.3.0
Django==4.2.1
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
docutils==0.19
dotty-dict==1.3.1
fastapi==0.95.1
filelock==3.12.2
firebase-admin==6.1.0
flatbuffers==23.3.3
fonttools==4.39.3
future==0.18.3
gast==0.4.0
gcloud==0.17.0
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.53.0
# gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.3.0
iniconfig==2.0.0
invoke==1.7.3
jaraco.classes==3.2.3
jax==0.4.8
joblib==1.2.0
jws==0.1.3
keras==2.12.0
keyring==23.13.1
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
more-itertools==9.1.0
msgpack==1.0.5
# nltk==3.8.1
# numpy==1.23.5
oauth2client==3.0.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.2
Pillow==9.5.0
pkginfo==1.9.6
pluggy==1.0.0
postgrest==0.10.6
postgrest-py==0.4.0
proto-plus==1.22.2
protobuf==4.22.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pyparsing==3.0.9
pypng==0.20220715.0
Pyrebase==3.0.27
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
python-gitlab==3.14.0
python-jwt==2.0.1
python-semantic-release==7.33.2
pytz==2023.3
qrcode==7.4.2
readme-renderer==37.3
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
rfc3986==1.5.0
rsa==4.9
# scipy==1.10.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
# SpeechRecognition==3.10.0
# sqlparse==0.4.4
starlette==0.26.1
storage3==0.5.2
StrEnum==0.4.10
# supabase==1.0.3
# supabase-py==0.0.2
# supafunc==0.2.2
# tensorboard==2.12.2
# tensorboard-data-server==0.7.0
# tensorboard-plugin-wit==1.8.1
# tensorflow==2.12.0
# tensorflow-estimator==2.12.0
# tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
# tflearn==0.5.0
tldextract==3.4.4
toml==0.10.2
tomlkit==0.11.7
tqdm==4.65.0
twine==3.8.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
# uvicorn==0.22.0
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
